refactor(extensions): move ContentTransformer to src/extensions/content/, rename ext-mdx → ext-content-mdx
- src/extensions/content/{content-transformer,index}.ts
- Barrel separates 3 type aliases (CompilationMode, CompilationTarget, ContentPlugin)
  from 3 interfaces (ContentCompileOptions, ContentRuntimeBundle, ContentTransformer)
  into distinct blocks
- src/extensions/interfaces/index.ts shim re-exports from ../content/
- extensions/ext-content-mdx (@veryfront/ext-content-mdx) maps veryfront/extensions/content
- Root deno.json workspace + ./extensions/content export + imports alias
- Consumer stragglers (builtin-extensions.ts, ensure-content-transformer.ts,
  content-transformer-setup.ts, recommendations.ts, transforms/plugins/*, tests) updated
