+++ extensions/ext-content-mdx/src/compiler/mdx-compile.ts
@@ -1,6 +1,6 @@
 import { compile } from "@mdx-js/mdx";
 import type { Pluggable } from "unified";
-import type { ContentCompileOptions, ContentRuntimeBundle } from "veryfront/extensions/interfaces";
+import type { ContentCompileOptions, ContentRuntimeBundle } from "veryfront/extensions/content";
 import { extractFrontmatter } from "veryfront/transforms/frontmatter";
 import { rewriteBodyImports, rewriteCompiledImports } from "veryfront/transforms/import-rewriter";
 import { getRehypePlugins, getRemarkPlugins } from "../plugins/plugin-loader.ts";
