+++ extensions/ext-content-mdx/src/plugins/rehype-mermaid.ts
