+++ extensions/ext-content-mdx/src/plugins/remark-mdx-utils.ts
