+++ src/extensions/content/index.ts
@@ -1,0 +1,19 @@
+/**
+ * Content category barrel — content transformer (MDX/Markdown) contract.
+ *
+ * @module extensions/content
+ */
+
+// Type aliases (unions / shape aliases)
+export type {
+  CompilationMode,
+  CompilationTarget,
+  ContentPlugin,
+} from "./content-transformer.ts";
+
+// Interfaces
+export type {
+  ContentCompileOptions,
+  ContentRuntimeBundle,
+  ContentTransformer,
+} from "./content-transformer.ts";
